my_site/requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.5.0
argon2-cffi==23.1.0
pillow==11.0.0
Faker==33.3.0
djangorestframework==3.15.2
django-filter==24.3
flake8==6.0.0
flake8-docstrings==1.7.0
Flake8-pyproject==1.2.3
flake8-django==1.4
mypy==1.14.1
isort==5.13.2
django-stubs==5.1.1
docutils==0.21.2
drf-spectacular==0.28.0
django-debug-toolbar==5.0.1
sentry-sdk==2.20.0
pydantic==2.10.6
pydantic-settings==2.7.1
